requirements.txt
requests
pandas
numpy
python-dotenv
scikit-learn
